+++ R/aaa.R
@@ -7,7 +7,7 @@ globalVariables(c(".","n",".resid","null.deviance","r_environment",".fitted","Pr
 #' @name radiant.model
 #' @docType package
 #' @import radiant.data shiny ggplot2
-#' @importFrom gridExtra grid.arrange arrangeGrob
+#' @importFrom gridExtra grid.arrange 
 #' @importFrom dplyr arrange select select_ filter mutate mutate_ mutate_each_ funs group_by group_by_ summarise summarise_ summarise_each summarise_each_ slice as_data_frame bind_cols bind_rows desc first last min_rank data_frame inner_join arrange_ ungroup
 #' @importFrom magrittr %>% %<>% %T>% set_colnames set_rownames set_names extract2
 #' @importFrom tidyr spread_ gather_ gather
@@ -64,4 +64,13 @@ NULL
 #' @name dvd
 #' @usage data(dvd)
 #' @format A data frame with 20,000 rows and 4 variables
+NULL
+
+#' Movie ratings
+#' @details Use collaborative filtering to create recommendations based on ratings from existing users. Description provided in attr(df,"description")
+#' @docType data
+#' @keywords datasets
+#' @name cf
+#' @usage data(cf)
+#' @format A data frame with 110 rows and 4 variables
 NULL